refactor(scripts): split AST-op types and extract bundle-installer
Zero-behavior-change reorganization of the scaffolding tooling internals.

- Extract the AST-operation type system out of integration-bundles.ts into a new
  ast-operation-types.ts (types only, no runtime) — cuts the circular-import risk
  of ast-transforms.ts importing op types from the bundle-data file. integration-
  bundles.ts re-exports the types for back-compat; ast-transforms imports from the
  new file directly.
- Extract the shared install sequence and the duplicated 'strip absent-integration
  wiring' block out of the executable satus.ts into a new bundle-installer.ts
  (installBundle + stripAbsentIntegrationWiring, one definition each). satus.ts keeps
  only arg-parsing / list / add orchestration; setup-project.ts imports the install
  primitives from bundle-installer.ts instead of from the executable.

No logic, control flow, signatures, or output changed — verified by the full suite
(337 tests, test:setup 35) and an independent Codex behavior-preservation pass. (#236)
